add lambda use to bench/durationcollector

+++ src/libs/antares/benchmarking/DurationCollector.cpp
@@ -24,7 +24,7 @@
 
 namespace Benchmarking {
 
-void DurationCollector::addDuration(const std::string& name, int64_t duration)
+void DurationCollector::addDuration(const std::string& name, long duration)
 {
     const std::lock_guard lock(mutex_);
     duration_items_[name].push_back(duration);
@@ -34,7 +34,7 @@ void DurationCollector::toFileContent(FileContent& file_content)
 {
     for (const auto& [name, durations] : duration_items_)
     {
-        const int64_t duration_sum = accumulate(durations.begin(), durations.end(), (int64_t)0);
+        const long duration_sum = accumulate(durations.begin(), durations.end(), 0);
 
         file_content.addDurationItem(name, (unsigned int)duration_sum, (int)durations.size());
     }
